requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
immutabledict==4.2.1
numpy==2.0.2
ortools==9.12.4544
pandas==2.2.3
protobuf==5.29.4
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.41
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
